pyproject.toml
[tool.poetry]
name = "nft-sets"
version = "0.1.0"
description = "Generate sets of images from a collection of layer files"
authors = ["DJ Stomp & Shadowdrums7331 <85457381+djstompzone@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.3.0"

[tool.poetry.group.dev.dependencies]
poetry = "^1.8.2"
pytest = "^8.2.0"
twine = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
